backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyjwt==2.8.0
openai==1.3.5
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
werkzeug==3.0.1
